+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Fix a line break issue when copying multiple messages.
 - Improve Github previews and add support for issues & pull requests too.
+- The home button can now be middle-clicked or `ctrl` / `cmd` clicked to open the homepage in a new tab / window.
 - Fix an issue with the `/mods` command not displaying its result.
 - Fix a context menu issue for moderators only always opening the timeout submenu by default.
 - Fix an auto-completion related issue that could lead to the first character being at the end ([#1](https://github.com/HiDeoo/YaTA/issues/1)).